Fix weather block and improve redis wrapper
