+++ idunn/blocks/environment.py
@@ -8,7 +8,7 @@ from idunn import settings
 from idunn.api.kuzzle import kuzzle_client
 from .base import BaseBlock
 
-from idunn.utils.redis import get_redis_pool
+from idunn.utils.redis import get_redis_pool, RedisWrapper
 
 
 logger = logging.getLogger(__name__)
@@ -105,67 +105,6 @@ class Weather(BaseBlock):
 
         return cls(**weather)
 
-    @classmethod
-    def set_value(cls, key, json_result):
-        try:
-            cls._connection.set(key, json_result, ex=cls._expire)
-        except RedisError:
-            prometheus.exception("RedisError")
-            logging.exception("Got a RedisError")
-
-    @classmethod
-    def get_value(cls, key):
-        try:
-            value_stored = cls._connection.get(key)
-            return value_stored
-        except RedisError as exc:
-            prometheus.exception("RedisError")
-            logging.exception("Got a RedisError")
-            raise CacheNotAvailable from exc
-
-    @classmethod
-    def init_cache(cls):
-        cls._expire = int(settings['WIKI_CACHE_TIMEOUT'])  # seconds
-        redis_db = settings['WIKI_CACHE_REDIS_DB']
-        try:
-            redis_pool = get_redis_pool(db=redis_db)
-        except RedisNotConfigured:
-            logger.warning("No Redis URL has been set for caching", exc_info=True)
-            cls._connection = DISABLED_STATE
-        else:
-            cls._connection = Redis(connection_pool=redis_pool)
-
-    @classmethod
-    def cache_it(cls, key, f):
-        """
-        Takes function f and put its result in a redis cache.
-        It requires a prefix string to identify the name
-        of the function cached.
-        """
-        if cls._connection is None:
-            cls.init_cache()
-
-        def with_cache(*args, **kwargs):
-            """
-            If the redis is up we use the cache,
-            otherwise we bypass it
-            """
-            if cls._connection is not DISABLED_STATE:
-                try:
-                    value_stored = cls.get_value(key)
-                except CacheNotAvailable:
-                    # Cache is not reachable: we don't want to execute 'f'
-                    # (and fetch wikipedia content, possibly very often)
-                    return None
-                if value_stored is not None:
-                    return json.loads(value_stored.decode('utf-8'))
-                result = f(*args, **kwargs)
-                json_result = json.dumps(result)
-                cls.set_value(key, json_result)
-                return result
-            return f(*args, **kwargs)
-        return with_cache
-
 
 def get_local_weather(coord):
     def inner(coord):
@@ -173,4 +112,5 @@ def get_local_weather(coord):
 
     if not weather_client.enabled:
         return None
-    wiki_poi_info = Weather.cache_it(coord, inner)(coord)
+    key = '{}_{}_{}'.format(Weather.BLOCK_TYPE, coord['lat'], coord['lon'])
+    return RedisWrapper.cache_it(key, inner)(coord)
