+++ tests/test_cache.py
@@ -5,6 +5,7 @@ from redis import Redis, RedisError
 from app import app, settings
 from starlette.testclient import TestClient
 from idunn.blocks.wikipedia import WikipediaCache
+from idunn.utils.redis import RedisWrapper
 from .test_wiki_ES import basket_ball_wiki_es
 from .test_rate_limiter import mock_wikipedia
 from functools import wraps
@@ -17,11 +18,11 @@ def cache_test_normal(redis):
     We define here settings specific to the
     test of the Wikipedia/Wikidata cache
     """
-    settings._settings["REDIS_URL"] = redis
-    WikipediaCache._connection = None
+    settings._settings['REDIS_URL'] = redis
+    RedisWrapper._connection = None
     yield
-    settings._settings["REDIS_URL"] = None
-    WikipediaCache._connection = None
+    settings._settings['REDIS_URL'] = None
+    RedisWrapper._connection = None
 
 
 def test_wikipedia_cache(cache_test_normal, mock_wikipedia):
