+++ tests/test_rate_limiter.py
@@ -5,6 +5,7 @@ from freezegun import freeze_time
 from app import app, settings
 from starlette.testclient import TestClient
 from idunn.blocks.wikipedia import WikipediaSession, WikipediaCache
+from idunn.utils.redis import RedisWrapper
 from .utils import override_settings
 
 from redis import RedisError
@@ -14,9 +15,10 @@ from functools import wraps
 
 @pytest.fixture(scope="function")
 def disable_wikipedia_cache():
-    WikipediaCache.disable()
+    RedisWrapper.disable()
     yield
-    WikipediaCache._connection = None
+    RedisWrapper._connection = None
+
 
 
 @pytest.fixture(scope="function")
@@ -51,7 +53,7 @@ def limiter_test_interruption(redis, disable_wikipedia_cache):
     WikipediaSession._rate_limiter = None
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture(scope='function')
 def mock_wikipedia(redis):
     with responses.RequestsMock(assert_all_requests_are_fired=False) as rsps:
         rsps.add(
@@ -231,6 +233,7 @@ def test_rate_limiter_with_redisError(limiter_test_interruption, mock_wikipedia,
     Here the redis is on so the answer should contain the wikipedia block
     """
     assert any(b["type"] == "wikipedia" for b in resp["blocks"][2].get("blocks"))
+
 
 
 @freeze_time("2018-06-14 8:30:00", tz_offset=2)
